fix compile error with cmake 4.3.1 in vs2022 and vs2026

+++ src/core/src/axe_core/system/SystemInfo.cpp
@@ -26,7 +26,7 @@ SystemInfo::OS SystemInfo::getOS() {
 }
 
 int SystemInfo::cpuCount() {
-	return ::std::thread::hardware_concurrency();
+	return static_cast<int>(::std::thread::hardware_concurrency());
 }
 
 } // namespace axe
